Finish project restructuring (now it actually compiles)
