+++ Neuro/Arrow/IArrowHandler.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 namespace Neuro.Arrow;
 
 // TODO: Is dependency injection really necessary for this?
-public interface IArrowHandler : IContextAcceptor
+public interface IArrowHandler : IContextAccepter
 {
     public LineRenderer Arrow { get; set; }
 }
