+++ Neuro/DependencyInjection/IContextProvider.cs
@@ -1,9 +1,9 @@
 ﻿using Neuro.Arrow;
+using Neuro.Minigames;
 using Neuro.Movement;
 using Neuro.Pathfinding;
 using Neuro.Recording;
 using Neuro.Tasks;
-using Neuro.Visibility;
 using Neuro.Vision;
 
 namespace Neuro.DependencyInjection;
@@ -11,10 +11,10 @@ namespace Neuro.DependencyInjection;
 public interface IContextProvider
 {
     public IArrowHandler ArrowHandler { get; }
+    public IMinigamesHandler MinigamesHandler { get; }
     public IMovementHandler MovementHandler { get; }
     public IPathfindingHandler PathfindingHandler { get; }
     public IRecordingHandler RecordingHandler { get; }
     public ITasksHandler TasksHandler { get; }
-    public IVisibilityHandler VisibilityHandler { get; }
     public IVisionHandler VisionHandler { get; }
 }
