+++ Neuro/Minigames/IMinigamesHandler.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections;
+using Neuro.DependencyInjection;
+
+namespace Neuro.Minigames;
+
+public interface IMinigamesHandler : IContextAccepter
+{
+    public IEnumerator CompleteMinigame(PlayerTask task, Minigame minigame);
+}
