+++ Neuro/Recording/DataStructures/MyVector2.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using JetBrains.Annotations;
 using UnityEngine;
 
 namespace Neuro.Recording.DataStructures;
