--- Neuro/Visibility/IVisibilityHandler.cs
@@ -1,7 +1,0 @@
-﻿using Neuro.DependencyInjection;
-
-namespace Neuro.Visibility;
-
-public interface IVisibilityHandler : IContextAcceptor
-{
-}
